Utilisation d'un composant CardAmenagement pour l'affichage des aménagements
