+++ frontend/src/controls/TabsContent/TabDecisionEtab.tsx
@@ -7,40 +7,28 @@
  * @author Julien Lemonnier <julien.lemonnier@u-bordeaux.fr>
  */
 
-import {
-   getAmenagementsDecision,
-   getDomaineAmenagement,
-} from "../../lib/amenagements";
+import { getAmenagementsDecision } from "../../lib/amenagements";
 import React, { useMemo } from "react";
-import { IAmenagement, ITypeAmenagement } from "../../api/ApiTypeHelpers";
+import { ITypeAmenagement } from "../../api/ApiTypeHelpers";
 import { useApi } from "../../context/api/ApiProvider";
 import {
    PREFETCH_CATEGORIES_AMENAGEMENTS,
    PREFETCH_TYPES_AMENAGEMENTS,
 } from "../../api/ApiPrefetchHelpers";
+import { CardAmenagement } from "../Card/CardAmenagement";
 import { NB_MAX_ITEMS_PER_PAGE } from "../../constants";
-import { Button, Card, Col, Empty, Flex, Row, Space, Tag, Tooltip, Typography } from "antd";
-import {
-   ArrowRightOutlined,
-   CalendarOutlined,
-   CommentOutlined,
-   EditOutlined,
-   HarmonyOSOutlined,
-} from "@ant-design/icons";
-import { getLibellePeriode } from "../../utils/dates";
-import SuiviAmenagementItem from "../Items/SuiviAmenagementItem";
+import { Empty, Flex, Row, Typography } from "antd";
+
 import { useSearchParams } from "react-router-dom";
 import useBreakpoint from "antd/es/grid/hooks/useBreakpoint";
-import { EllipsisParagraph } from "../Typography/EllipsisParagraph";
 import { BoutonDecisionEtab } from "./BoutonDecisionEtab";
-import { DOMAINES_AMENAGEMENTS_INFOS } from "../../lib/amenagements";
+import { getDomaineAmenagement } from "../../lib/amenagements";
 
 export function TabDecisionEtab(props: {
    utilisateurId: string;
 }) {
    const screens = useBreakpoint();
    const [searchParams] = useSearchParams();
-   const [editedAmenagement, setEditedAmenagement] = React.useState<IAmenagement>();
    const { data: typesAmenagements } = useApi().useGetCollection(PREFETCH_TYPES_AMENAGEMENTS);
    const { data: categoriesAmenagements } = useApi().useGetCollection(
       PREFETCH_CATEGORIES_AMENAGEMENTS,
@@ -63,11 +51,8 @@ export function TabDecisionEtab(props: {
    }, [
       amenagements?.items,
       categoriesAmenagements?.items,
-      DOMAINES_AMENAGEMENTS_INFOS.examen,
       typesAmenagements?.items,
    ]);
-
-   const highlightAmenagement = searchParams.get("amenagement");
 
    return (
       <>
@@ -76,7 +61,7 @@ export function TabDecisionEtab(props: {
                Décision d'établissement
             </Typography.Title>
             <div className={`text-right${!screens.lg ? " mt-2" : ""}`}>
-                  <BoutonDecisionEtab utilisateurId={props.utilisateurId} />
+               <BoutonDecisionEtab utilisateurId={props.utilisateurId} />
             </div>
          </Flex>
 
@@ -91,90 +76,12 @@ export function TabDecisionEtab(props: {
                         {c.typeAmenagements.map((ta) => (
                            <span key={ta["@id"]} style={{ width: "100%", display: "contents" }}>
                               {ta.amenagements.map((a) => (
-                                 <Col key={a["@id"]} xs={24} sm={24} md={24} lg={12} xl={8} xxl={6}>
-                                    <Card
-                                       style={{ borderColor: "#e0e0e0" }}
-                                       styles={
-                                          screens.lg ? { body: { minHeight: 195 } } : undefined
-                                       }
-                                       className={`${
-                                          highlightAmenagement === a["@id"]
-                                             ? "highlightAmenagement"
-                                             : ""
-                                       } bg-${getDomaineAmenagement(ta as ITypeAmenagement)?.couleur}-xxlight`}
-                                    >
-                                       <Card.Meta
-                                          title={
-                                             <Flex
-                                                wrap="wrap"
-                                                className="w-100"
-                                                justify="space-between"
-                                                align="start"
-                                             >
-                                                <div>
-                                                   <h5 className="mt-0 mb-0 fs-11">{ta.libelle}</h5>
-                                                   <span
-                                                      className={`text-${getDomaineAmenagement(ta as ITypeAmenagement)?.couleur}-dark fs-09`}
-                                                   >
-                                                      <ArrowRightOutlined className="mr-1" />
-                                                      {c.libelle}
-                                                   </span>
-                                                </div>
-                                                <Button
-                                                   className="mr-0 pr-0"
-                                                   type="text"
-                                                   icon={<EditOutlined aria-label="Menu" />}
-                                                   onClick={() => setEditedAmenagement(a)}
-                                                />
-                                             </Flex>
-                                          }
-                                          description={
-                                             <Space direction="vertical" className="text-text">
-                                                {a.debut || a.fin ? (
-                                                   <Space align="start" size={[8, 2]} wrap>
-                                                      <CalendarOutlined />
-                                                      <Tag>
-                                                         {getLibellePeriode(a.debut, a.fin, "MMM")}
-                                                      </Tag>
-                                                      <Space size={0}>
-                                                         {a.semestre1 && (
-                                                            <Tooltip title="Semestre 1">
-                                                               <Tag>S1</Tag>
-                                                            </Tooltip>
-                                                         )}
-                                                         {a.semestre2 && (
-                                                            <Tooltip title="Semestre 2">
-                                                               <Tag>S2</Tag>
-                                                            </Tooltip>
-                                                         )}
-                                                      </Space>
-                                                   </Space>
-                                                ) : null}
-                                                {a.commentaire && a.commentaire.length > 0 && (
-                                                   <Space align="start" size={12}>
-                                                      <CommentOutlined />
-                                                      <EllipsisParagraph
-                                                         content={a.commentaire}
-                                                         className="light mb-0 fs-09"
-                                                         type="secondary"
-                                                      />
-                                                   </Space>
-                                                )}
-                                                {a.suivi ? (
-                                                   <Space align="start" size={12}>
-                                                      <HarmonyOSOutlined />
-                                                      <SuiviAmenagementItem
-                                                         suiviId={a.suivi}
-                                                         className="float-right"
-                                                         couleur={getDomaineAmenagement(ta as ITypeAmenagement)?.couleur}
-                                                      />
-                                                   </Space>
-                                                ) : null}
-                                             </Space>
-                                          }
-                                       />
-                                    </Card>
-                                 </Col>
+                                 <CardAmenagement
+                                    couleur={getDomaineAmenagement(ta as ITypeAmenagement)?.couleur} 
+                                    categorie={c.libelle!} 
+                                    amenagement={a} 
+                                    type={ta.libelle}>
+                                 </CardAmenagement>
                               ))}
                            </span>
                         ))}
@@ -189,4 +96,3 @@ export function TabDecisionEtab(props: {
       </>
    );
 }
-
